\name{fisherInformation-methods}
\docType{methods}
\alias{fisherInformation-methods}
\alias{fisherInformation,aclinicalProteomicsData-method}
\title{Methods for Function fisherInformation}
\description{
 Methods for function \code{fisherInformation} are
 defined with class "aclinicalProteomicsData" in the signature.
}
\section{Methods}{
\describe{

\item{Data = "aclinicalProteomicsData"}{This is a class object for the mass spectrometry data sets, which 
are in the same format as the raw data from the Biomarkers wizard software. 
It has slots of matrices of raw mass spectrometry and phenotypic data sets, a 
character variable for the classes of all the covariates in the phenotypic data 
matrix, a character variable for the covariates of interest, and numeric value for 
the number of peaks of interest. See also \code{aclinicalProteomicsData-class}.
}
}}
\keyword{methods}
